// examples/react/src/skins/minimal/types.ts
export interface MinimalSkinStyles {
  readonly MediaContainer: string;
  readonly Overlay: string;
  readonly Controls: string;
  readonly Icon: string;
  readonly Button: string;
  readonly ButtonGroup: string;
  readonly IconButton: string;
  readonly PlayButton: string;
  readonly PlayIcon: string;
  readonly PauseIcon: string;
  readonly MuteButton: string;
  readonly VolumeHighIcon: string;
  readonly VolumeLowIcon: string;
  readonly VolumeOffIcon: string;
  readonly FullscreenButton: string;
  readonly FullscreenEnterIcon: string;
  readonly FullscreenExitIcon: string;
  readonly TimeSliderRoot: string;
  readonly TimeSliderThumb: string;
  readonly TimeDisplay: string;
  readonly SliderRoot: string;
  readonly SliderTrack: string;
  readonly SliderProgress: string;
  readonly SliderPointer: string;
  readonly SliderThumb: string;
}